@@ -1,6 +1,8 @@
+using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
+using System.Text.Json;
 using DotVVM.Core.Common;
 using DotVVM.Framework.Api.Swashbuckle.AspNetCore.Filters;
 using DotVVM.Samples.BasicSamples.Api.AspNetCore.Controllers;
@@ -14,8 +16,11 @@ using Microsoft.AspNetCore.Mvc.Internal;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.ObjectPool;
 using Microsoft.Extensions.Options;
+using Microsoft.OpenApi.Any;
+using Microsoft.OpenApi.Models;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Newtonsoft.Json;
+using NSwag.SwaggerGeneration.WebApi.Infrastructure;
 using Swashbuckle.AspNetCore.Swagger;
 using Swashbuckle.AspNetCore.SwaggerGen;
 
@@ -24,7 +29,7 @@ namespace DotVVM.Framework.Api.Swashbuckle.AspNetCore.Tests
     [TestClass]
     public class SwaggerFileGenerationTests
     {
-        private static SwaggerDocument document;
+        private static OpenApiDocument document;
 
         [ClassInitialize]
         public static void Init(TestContext testContext)
@@ -48,7 +53,7 @@ namespace DotVVM.Framework.Api.Swashbuckle.AspNetCore.Tests
                 },
                 SwaggerDocs =
                 {
-                    { "v1", new Info() { Title = "Test API", Version = "v1" }}
+                    { "v1", new OpenApiInfo() { Title = "Test API", Version = "v1" }}
                 }
             };
 
@@ -64,62 +69,83 @@ namespace DotVVM.Framework.Api.Swashbuckle.AspNetCore.Tests
             var serviceProvider = serviceCollection.BuildServiceProvider();
             var apiDescriptionGroupCollectionProvider = serviceProvider.GetRequiredService<IApiDescriptionGroupCollectionProvider>();
 
-            var schemaSettings = new SchemaRegistryOptions()
+            string GetCustomSchemaId(Type modelType)
             {
+                if (!modelType.IsConstructedGenericType) return modelType.Name.Replace("[]", "Array");
+
+                var generics = modelType.GetGenericArguments()
+                    .Select(genericArg => GetCustomSchemaId(genericArg))
+                    .Aggregate((previous, current) => previous + current);
+
+                return $"{modelType.Name.Split('`').First()}[{generics}]";
+            }
+
+            var schemaGeneratorOptions = new SchemaGeneratorOptions() {
                 SchemaFilters =
                 {
                     new AddTypeToModelSchemaFilter()
-                }
+                },
+                SchemaIdSelector = type => GetCustomSchemaId(type)
             };
 
-            var schemaRegistryFactory = new SchemaRegistryFactory(new JsonSerializerSettings(), schemaSettings);
-            var generator = new SwaggerGenerator(apiDescriptionGroupCollectionProvider, schemaRegistryFactory, options);
-            document = generator.GetSwagger("v1");
+            var schemaGenerator = new SchemaGenerator(schemaGeneratorOptions, new JsonSerializerDataContractResolver(new JsonSerializerOptions()));
+            var swaggerGenerator = new SwaggerGenerator(options, apiDescriptionGroupCollectionProvider, schemaGenerator);
+
+            document = swaggerGenerator.GetSwagger("v1");
         }
 
         [TestMethod]
         public void Swashbuckle_AspNetCore_NameAndKnownTypeAnnotation_Added()
         {
-            var definition = document.Definitions["IPagingOptions"];
+            var definition = document.Components.Schemas["IPagingOptions"];
 
-            Assert.AreEqual("DotVVM.Framework.Controls.IPagingOptions", definition.Extensions[ApiConstants.DotvvmKnownTypeKey]);
-            Assert.AreEqual("PageIndex", definition.Properties["PageIndex"].Extensions[ApiConstants.DotvvmNameKey]);
-            Assert.AreEqual("PageSize", definition.Properties["PageSize"].Extensions[ApiConstants.DotvvmNameKey]);
+            var knownTypeKeyWrapper = definition.Extensions[ApiConstants.DotvvmKnownTypeKey] as OpenApiString;
+            Assert.AreEqual("DotVVM.Framework.Controls.IPagingOptions", knownTypeKeyWrapper.Value);
+
+            var nameKeyWrapper = definition.Properties["PageIndex"].Extensions[ApiConstants.DotvvmNameKey] as OpenApiString;
+            Assert.AreEqual("PageIndex", nameKeyWrapper.Value);
+
+            nameKeyWrapper = definition.Properties["PageSize"].Extensions[ApiConstants.DotvvmNameKey] as OpenApiString;
+            Assert.AreEqual("PageSize", nameKeyWrapper.Value);
         }
 
         [TestMethod]
         public void Swashbuckle_AspNetCore_GenericNameAndKnownTypeAnnotation_Added()
         {
-            var definition = document.Definitions["GridViewDataSet[Company[String]]"];
+            var definition = document.Components.Schemas["GridViewDataSet[Company[String]]"];
 
-            Assert.AreEqual("DotVVM.Framework.Controls.GridViewDataSet<Company[String]>",
-                definition.Extensions[ApiConstants.DotvvmKnownTypeKey]);
+            var knownTypeKeyWrapper = definition.Extensions[ApiConstants.DotvvmKnownTypeKey] as OpenApiString;
+            Assert.AreEqual("DotVVM.Framework.Controls.GridViewDataSet<Company[String]>", knownTypeKeyWrapper.Value);
 
-            Assert.AreEqual("IsRefreshRequired", definition.Properties["IsRefreshRequired"].Extensions[ApiConstants.DotvvmNameKey]);
-            Assert.AreEqual("Items", definition.Properties["Items"].Extensions[ApiConstants.DotvvmNameKey]);
+            var nameKeyWrapper = definition.Properties["IsRefreshRequired"].Extensions[ApiConstants.DotvvmNameKey] as OpenApiString;
+            Assert.AreEqual("IsRefreshRequired", nameKeyWrapper.Value);
+
+            nameKeyWrapper = definition.Properties["Items"].Extensions[ApiConstants.DotvvmNameKey] as OpenApiString;
+            Assert.AreEqual("Items", nameKeyWrapper.Value);
         }
 
         [TestMethod]
         public void Swashbuckle_AspNetCore_GenericKnownTypeAnnotation_FullNameParameter()
         {
-            var definition = document.Definitions["Company[String]"];
+            var definition = document.Components.Schemas["Company[String]"];
 
-            Assert.AreEqual("DotVVM.Samples.BasicSamples.Api.Common.Model.Company<System.String>",
-                definition.Extensions[ApiConstants.DotvvmKnownTypeKey]);
+            var wrappedKnownTypeKey = definition.Extensions[ApiConstants.DotvvmKnownTypeKey] as OpenApiString;
+            Assert.AreEqual("DotVVM.Samples.BasicSamples.Api.Common.Model.Company<System.String>", wrappedKnownTypeKey.Value);
         }
 
         [TestMethod]
         public void Swashbuckle_AspNetCore_MixedGenericKnownTypeAnnotation_Exist()
         {
-            var definition = document.Definitions["GridViewDataSet[Company[Boolean]]"];
+            var definition = document.Components.Schemas["GridViewDataSet[Company[Boolean]]"];
 
-            Assert.AreEqual("DotVVM.Framework.Controls.GridViewDataSet<Company[Boolean]>", definition.Extensions[ApiConstants.DotvvmKnownTypeKey]);
+            var wrappedKnownTypeKey = definition.Extensions[ApiConstants.DotvvmKnownTypeKey] as OpenApiString;
+            Assert.AreEqual("DotVVM.Framework.Controls.GridViewDataSet<Company[Boolean]>", wrappedKnownTypeKey.Value);
         }
 
         [TestMethod]
         public void Swashbuckle_AspNetCore_KnownTypeAnnotation_NotAdded()
         {
-            var definition = document.Definitions["Order"];
+            var definition = document.Components.Schemas["Order"];
 
             Assert.IsFalse(definition.Extensions.ContainsKey(ApiConstants.DotvvmKnownTypeKey));
         }
@@ -127,25 +153,28 @@ namespace DotVVM.Framework.Api.Swashbuckle.AspNetCore.Tests
         [TestMethod]
         public void Swashbuckle_AspNetCore_WrapperTypeAnnotation_OneArgument()
         {
-            var operation1 = document.Paths["/api/Companies/sorted"].Get;
+            var operation1 = document.Paths["/api/Companies/sorted"].Operations[OperationType.Get];
             var param1 = operation1.Parameters.Single(p => p.Name == "sortingOptions.SortDescending");
-            Assert.AreEqual("DotVVM.Framework.Controls.ISortingOptions, DotVVM.Core", param1.Extensions["x-dotvvm-wrapperType"]);
+            var param1WrapperType = param1.Extensions["x-dotvvm-wrapperType"] as OpenApiPrimitive<string>;
+            Assert.AreEqual("DotVVM.Framework.Controls.ISortingOptions, DotVVM.Core", param1WrapperType.Value);
         }
 
         [TestMethod]
         public void Swashbuckle_AspNetCore_WrapperTypeAnnotation_TwoArguments()
         {
-            var operation = document.Paths["/api/Companies/sortedandpaged"].Get;
+            var operation = document.Paths["/api/Companies/sortedandpaged"].Operations[OperationType.Get];
             var param1 = operation.Parameters.Single(p => p.Name == "sortingOptions.SortDescending");
+            var param1WrapperType = param1.Extensions["x-dotvvm-wrapperType"] as OpenApiPrimitive<string>;
             var param2 = operation.Parameters.Single(p => p.Name == "pagingOptions.PageSize");
-            Assert.AreEqual("DotVVM.Framework.Controls.ISortingOptions, DotVVM.Core", param1.Extensions["x-dotvvm-wrapperType"]);
-            Assert.AreEqual("DotVVM.Framework.Controls.IPagingOptions, DotVVM.Core", param2.Extensions["x-dotvvm-wrapperType"]);
+            var param2WrapperType = param2.Extensions["x-dotvvm-wrapperType"] as OpenApiPrimitive<string>;
+            Assert.AreEqual("DotVVM.Framework.Controls.ISortingOptions, DotVVM.Core", param1WrapperType.Value);
+            Assert.AreEqual("DotVVM.Framework.Controls.IPagingOptions, DotVVM.Core", param2WrapperType.Value);
         }
 
         [TestMethod]
         public void Swashbuckle_AspNetCore_WrapperTypeAnnotation_NotAdded()
         {
-            var operation = document.Paths["/api/Orders/{orderId}"].Get;
+            var operation = document.Paths["/api/Orders/{orderId}"].Operations[OperationType.Get];
             var param = operation.Parameters.Single(p => p.Name == "orderId");
             Assert.IsFalse(param.Extensions.ContainsKey("x-dotvvm-wrapperType"));
         }
