@@ -1,6 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using Microsoft.AspNetCore.Mvc.ModelBinding.Metadata;
+using Microsoft.OpenApi.Models;
 using Swashbuckle.AspNetCore.Swagger;
 using Swashbuckle.AspNetCore.SwaggerGen;
 
@@ -8,14 +10,14 @@ namespace DotVVM.Framework.Api.Swashbuckle.AspNetCore.Filters
 {
     public class RemoveReadOnlyFromUriParametersOperationFilter : IOperationFilter
     {
-        public void Apply(Operation operation, OperationFilterContext context)
+        public void Apply(OpenApiOperation operation, OperationFilterContext context)
         {
             if (operation.Parameters != null)
             {
                 foreach (var param in operation.Parameters.ToList())
                 {
                     var description = context.ApiDescription.ParameterDescriptions.SingleOrDefault(p => p.Name == param.Name);
-                    if (description?.ModelMetadata.IsReadOnly == true)
+                    if (description?.ModelMetadata.IsReadOnly == true && description?.ModelMetadata.MetadataKind == ModelMetadataKind.Property)
                     {
                         operation.Parameters.Remove(param);
                     }
